+++ migrations/m150614_103145_update_social_account_table.php
@@ -1,25 +1,24 @@
 <?php
 
 use yii\db\Query;
-use yii\db\Schema;
-use yii\db\Migration;
+use dektrium\user\migrations\Migration;
 
 class m150614_103145_update_social_account_table extends Migration
 {
     public function up()
     {
-        $this->addColumn('{{%social_account}}', 'code', Schema::TYPE_STRING . '(32)');
-        $this->addColumn('{{%social_account}}', 'created_at', Schema::TYPE_INTEGER);
-        $this->addColumn('{{%social_account}}', 'email', Schema::TYPE_STRING);
-        $this->addColumn('{{%social_account}}', 'username', Schema::TYPE_STRING);
-        $this->createIndex('account_unique_code', '{{%social_account}}', 'code', true);
+        $this->addColumn('{{%social_account}}', 'code', $this->string(32)->null());
+        $this->addColumn('{{%social_account}}', 'created_at', $this->integer()->null());
+        $this->addColumn('{{%social_account}}', 'email', $this->string()->null());
+        $this->addColumn('{{%social_account}}', 'username', $this->string()->null());
+        $this->createIndex('{{%account_unique_code}}', '{{%social_account}}', 'code', true);
 
-        $accounts = (new Query())->from('{{%social_account}}')->select('id')->all();
+        $accounts = (new Query())->from('{{%social_account}}')->select('id')->all($this->db);
 
-        $transaction = Yii::$app->db->beginTransaction();
+        $transaction = $this->db->beginTransaction();
         try {
             foreach ($accounts as $account) {
-                Yii::$app->db->createCommand()->update('{{%social_account}}', [
+                $this->db->createCommand()->update('{{%social_account}}', [
                     'created_at' => time(),
                 ], 'id = ' . $account['id'])->execute();
             }
@@ -32,7 +31,7 @@ class m150614_103145_update_social_account_table extends Migration
 
     public function down()
     {
-        $this->dropIndex('account_unique_code', '{{%social_account}}');
+        $this->dropIndex('{{%account_unique_code}}', '{{%social_account}}');
         $this->dropColumn('{{%social_account}}', 'email');
         $this->dropColumn('{{%social_account}}', 'username');
         $this->dropColumn('{{%social_account}}', 'code');