+++ models/LoginForm.php
@@ -13,9 +13,11 @@ namespace dektrium\user\models;
 
 use dektrium\user\Finder;
 use dektrium\user\helpers\Password;
+use dektrium\user\traits\ModuleTrait;
+use yii\helpers\ArrayHelper;
+use yii\helpers\Html;
 use Yii;
 use yii\base\Model;
-use dektrium\user\traits\ModuleTrait;
 
 /**
  * LoginForm get user's login and password, validates them and logs the user in. If user has been blocked, it adds
@@ -52,6 +54,23 @@ class LoginForm extends Model
         parent::__construct($config);
     }
 
+    /**
+     * Gets all users to generate the dropdown list when in debug mode.
+     *
+     * @return array
+     */
+    public static function loginList()
+    {
+        /** @var \dektrium\user\Module $module */
+        $module = \Yii::$app->getModule('user');
+
+        $userModel = $module->modelMap['User'];
+
+        return ArrayHelper::map($userModel::find()->where(['blocked_at' => null])->all(), 'username', function ($user) {
+            return sprintf('%s (%s)', Html::encode($user->username), Html::encode($user->email));
+        });
+    }
+
     /** @inheritdoc */
     public function attributeLabels()
     {
@@ -65,17 +84,9 @@ class LoginForm extends Model
     /** @inheritdoc */
     public function rules()
     {
-        return [
-            'requiredFields' => [['login', 'password'], 'required'],
+        $rules = [
             'loginTrim' => ['login', 'trim'],
-            'passwordValidate' => [
-                'password',
-                function ($attribute) {
-                    if ($this->user === null || !Password::validate($this->password, $this->user->password_hash)) {
-                        $this->addError($attribute, Yii::t('user', 'Invalid login or password'));
-                    }
-                }
-            ],
+            'requiredFields' => [['login'], 'required'],
             'confirmationValidate' => [
                 'login',
                 function ($attribute) {
@@ -93,6 +104,34 @@ class LoginForm extends Model
             ],
             'rememberMe' => ['rememberMe', 'boolean'],
         ];
+
+        if (!$this->module->debug) {
+            $rules = array_merge($rules, [
+                'requiredFields' => [['login', 'password'], 'required'],
+                'passwordValidate' => [
+                    'password',
+                    function ($attribute) {
+                        if ($this->user === null || !Password::validate($this->password, $this->user->password_hash)) {
+                            $this->addError($attribute, Yii::t('user', 'Invalid login or password'));
+                        }
+                    }
+                ]
+            ]);
+        }
+
+        return $rules;
+    }
+
+    /**
+     * Validates if the hash of the given password is identical to the saved hash in the database.
+     * It will always succeed if the module is in DEBUG mode.
+     *
+     * @return void
+     */
+    public function validatePassword($attribute, $params)
+    {
+      if ($this->user === null || !Password::validate($this->password, $this->user->password_hash))
+        $this->addError($attribute, Yii::t('user', 'Invalid login or password'));
     }
 
     /**
@@ -102,12 +141,19 @@ class LoginForm extends Model
      */
     public function login()
     {
-        if ($this->validate()) {
-            return Yii::$app->getUser()->login($this->user, $this->rememberMe ? $this->module->rememberFor : 0);
-        } else {
-            return false;
+        if ($this->validate() && $this->user) {
+            $isLogged = Yii::$app->getUser()->login($this->user, $this->rememberMe ? $this->module->rememberFor : 0);
+
+            if ($isLogged) {
+                $this->user->updateAttributes(['last_login_at' => time()]);
+            }
+
+            return $isLogged;
         }
+
+        return false;
     }
+
 
     /** @inheritdoc */
     public function formName()