+++ views/security/login.php
@@ -10,13 +10,14 @@
  */
 
 use dektrium\user\widgets\Connect;
+use dektrium\user\models\LoginForm;
 use yii\helpers\Html;
 use yii\widgets\ActiveForm;
 
 /**
- * @var yii\web\View                   $this
+ * @var yii\web\View $this
  * @var dektrium\user\models\LoginForm $model
- * @var dektrium\user\Module           $module
+ * @var dektrium\user\Module $module
  */
 
 $this->title = Yii::t('user', 'Sign in');
@@ -33,43 +34,58 @@ $this->params['breadcrumbs'][] = $this->title;
             </div>
             <div class="panel-body">
                 <?php $form = ActiveForm::begin([
-                    'id'                     => 'login-form',
-                    'enableAjaxValidation'   => true,
+                    'id' => 'login-form',
+                    'enableAjaxValidation' => true,
                     'enableClientValidation' => false,
-                    'validateOnBlur'         => false,
-                    'validateOnType'         => false,
-                    'validateOnChange'       => false,
+                    'validateOnBlur' => false,
+                    'validateOnType' => false,
+                    'validateOnChange' => false,
                 ]) ?>
 
-                <?= $form->field(
-                    $model,
-                    'login',
-                    ['inputOptions' => ['autofocus' => 'autofocus', 'class' => 'form-control', 'tabindex' => '1']]
-                ) ?>
+                <?php if ($module->debug): ?>
+                    <?= $form->field($model, 'login', [
+                        'inputOptions' => [
+                            'autofocus' => 'autofocus',
+                            'class' => 'form-control',
+                            'tabindex' => '1']])->dropDownList(LoginForm::loginList());
+                    ?>
 
-                <?= $form
-                    ->field(
+                <?php else: ?>
+
+                    <?= $form->field($model, 'login',
+                        ['inputOptions' => ['autofocus' => 'autofocus', 'class' => 'form-control', 'tabindex' => '1']]
+                    );
+                    ?>
+
+                <?php endif ?>
+
+                <?php if ($module->debug): ?>
+                    <div class="alert alert-warning">
+                        <?= Yii::t('user', 'Password is not necessary because the module is in DEBUG mode.'); ?>
+                    </div>
+                <?php else: ?>
+                    <?= $form->field(
                         $model,
                         'password',
-                        ['inputOptions' => ['class' => 'form-control', 'tabindex' => '2']]
-                    )
-                    ->passwordInput()
-                    ->label(
-                        Yii::t('user', 'Password')
-                        .($module->enablePasswordRecovery ?
-                            ' (' . Html::a(
-                                Yii::t('user', 'Forgot password?'),
-                                ['/user/recovery/request'],
-                                ['tabindex' => '5']
-                            )
-                            . ')' : '')
-                    ) ?>
+                        ['inputOptions' => ['class' => 'form-control', 'tabindex' => '2']])
+                        ->passwordInput()
+                        ->label(
+                            Yii::t('user', 'Password')
+                            . ($module->enablePasswordRecovery ?
+                                ' (' . Html::a(
+                                    Yii::t('user', 'Forgot password?'),
+                                    ['/user/recovery/request'],
+                                    ['tabindex' => '5']
+                                )
+                                . ')' : '')
+                        ) ?>
+                <?php endif ?>
 
-                <?= $form->field($model, 'rememberMe')->checkbox(['tabindex' => '4']) ?>
+                <?= $form->field($model, 'rememberMe')->checkbox(['tabindex' => '3']) ?>
 
                 <?= Html::submitButton(
                     Yii::t('user', 'Sign in'),
-                    ['class' => 'btn btn-primary btn-block', 'tabindex' => '3']
+                    ['class' => 'btn btn-primary btn-block', 'tabindex' => '4']
                 ) ?>
 
                 <?php ActiveForm::end(); ?>