+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # CHANGELOG
 
+- Fix: Update last_login_at when successfully logged in over oauth(#899)
+
+## 0.9.12 [11 January 2017]
+
+- Fix: Fixed migrations (dmeroff)
+- Fix #832: Fixed admin view file (thyseus)
+- Enh #839: Order by most recent registration first (created_at DESC) by default (thyseus)
+
+## 0.9.11 [10 January 2017]
+
+- Fix: Fixed migrations for postgresql (dmeroff)
+- Enh #794: Allow fast login without password when in DEBUG mode (thyseus)
+- Enh #826: Auth action added to auth rules (faenir)
+- Enh #820: Added ability to switch back to latest user after being impersonated (thyseus)
+- Enh #774: Added last_login_at column to user table (thyseus)
+
+## 0.9.10 [8 January 2017]
+
+- Enh #767: Added support for MS SQL Server (tsdogs)
+- Enh #795: Added method `getAccountByProvider` to User model to get account model by provider name (dmeroff)
+- Fix #778: Migrations now use correct db component (dmeroff)
+- Fix #777: Rethrow exception on failed user creation or registration (dmeroff)
+- Enh #772, #791: Added ability to log into another user's accoun (thyseus)
+- Fix #761: Fixed EVENT_AFTER_CONFIRM not triggering on user creation (dmeroff)
+- Fix #757: Fixed tabindex order in security/login.php view (dmeroff)
+
 ## 0.9.9 [13 August 2016]
 
 - Enh #746: Changed AccessRule to check other roles/permissions (markux)